Time: 15 ms (19.24%), Space: 26.1 MB (81.48%) - LeetHub

+++ 1283-find-the-smallest-divisor-given-a-threshold/1283-find-the-smallest-divisor-given-a-threshold.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
-int sumofd(vector<int>& nums,int d){
-    int sum=0;
+long long sumofd(vector<int>& nums,int d){
+    long long sum=0;
     int n=nums.size();
     for(int i=0;i<n;i++){
         sum+=ceil((double)nums[i]/(double)d);
